cleaning common module moved includes to include/spine
