+++ include/spine/topic_factory.h
