+++ src/topic_factory.cc
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "topic_factory.h"
+#include "spine/topic_factory.h"
 
 #include "common.h"
 
@@ -43,7 +43,7 @@ std::unique_ptr<Topic> TopicFactory::for_target(std::unique_ptr<Target>&& target
     Topic* topic = Topic::default_instance().New();
     topic->set_allocated_id(topic_id);
     topic->set_allocated_target(target.release());
-    topic->set_allocated_context(copy_actor_context(*actor_context_));
+    topic->set_allocated_context(clone(*actor_context_));
 
     return std::unique_ptr<Topic>(topic);
 }
